Adaptation of the code to the new omnetpp 4.1 version. This code is compatible with version 4.0 and 4.1b4, but incompatible with 4.b1, 4.b2 and 4.b3

+++ src/networklayer/manetrouting/dymo_fau/DYMO.cc
@@ -680,7 +680,7 @@ void DYMO::sendDown(cPacket* apMsg, int destAddr) {
 	// keep statistics
 	totalPacketsSent++;
 	totalBytesSent+=apMsg->getByteLength();
-	if (LL_MANET_ROUTERS.getInt()==destAddr)
+	if (LL_MANET_ROUTERS.getInt()==(unsigned int)destAddr)
 	{
 		destAddr = IPAddress::ALLONES_ADDRESS.getInt();
 		sendToIp(apMsg, UDPPort, destAddr, UDPPort,1, SIMTIME_DBL(jitter),0);
